install as root when possible

+++ R/zzz_centos_base.R
@@ -1,6 +1,8 @@
 centos_requirements <- function() rpm_requirements()
 
 centos_install <- function(pkgs) {
+  if (root()) return(centos_install_root(pkgs))
+
   dir.create(temp <- tempfile("rspm_"))
   old <- setwd(temp)
   on.exit({ setwd(old); unlink(temp, recursive=TRUE, force=TRUE) })
@@ -10,6 +12,13 @@ centos_install <- function(pkgs) {
     cmd <- "dnf download --resolve"
   system(cmd, p(pkgs))
   rpm_install()
+}
+
+centos_install_root <- function(pkgs) {
+  cmd <- "yum"
+  if (Sys.which("dnf") != "")
+    cmd <- "dnf"
+  system(cmd, "-y install", p(pkgs))
 }
 
 centos_install_sysreqs <- function(libs) {
