+++ R/zzz_opensuse_base.R
@@ -1,12 +1,18 @@
 opensuse_requirements <- function() rpm_requirements()
 
 opensuse_install <- function(pkgs) {
+  if (root()) return(opensuse_install_root(pkgs))
+
   dir.create(temp <- tempfile("rspm_"))
   old <- setwd(temp)
   on.exit({ setwd(old); unlink(temp, recursive=TRUE, force=TRUE) })
 
   system("zypper --pkg-cache-dir . install -dy", p(pkgs))
   rpm_install()
+}
+
+opensuse_install_root <- function(pkgs) {
+  system("zypper install -y", p(pkgs))
 }
 
 opensuse_install_sysreqs <- function(libs) {
